requirements.txt
#langchain core
langchain
langchain-core

langchain-openai
openai

langchain-anthropic
langchain-google-genai
google-generativeai

langchain-huggingface
transformers
huggingface-hub

python-dotenv

numpy 
scikit-learn
pandas
pydantic
langchain-community
